# docs/swagger.yaml
basePath: /
definitions:
...
  handler.errorResponse:
    properties:
      code:
        example: ERR_INVALID_REQUEST
        type: string
      error:
        example: error description
        type: string
    type: object
...
  handler.loginResponse:
    properties:
      expire_date:
        example: "2025-01-01T00:00:00Z"
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  handler.messageResponse:
    properties:
      message:
        example: Operation successful
        type: string
    type: object
...
  handler.sendCodeRequest:
    properties:
      allow_fallback:
        example: true
        type: boolean
      device_uid:
        example: device-uuid-1234
        type: string
      provider:
        example: telegram
        type: string
      recipient:
        example: user@example.com
        type: string
    type: object
  handler.verifyCodeRequest:
    properties:
      code:
        example: "123456"
        type: string
      device_uid:
        example: device-uuid-1234
        type: string
      recipient:
        example: user@example.com
        type: string
    type: object
  handler.verifyLogin2FARequest:
    properties:
      code:
        example: "123456"
        type: string
      device_uid:
        example: device-uuid-1234
        type: string
      login:
        example: user@example.com
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: darkrain
    url: https://github.com/darkrain/auth-service
  description: Authentication and authorization microservice with JWT sessions, 2FA,
    and API key management.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Auth Service API
  version: "1.0"
paths:
...
  /auth/login/verify-2fa:
    post:
      consumes:
      - application/json
      description: Completes the login process by verifying a 2FA code. Returns a
        JWT token on success.
      parameters:
      - description: Login, 2FA code and device UID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.verifyLogin2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.loginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Verify 2FA code during login
      tags:
      - auth
...
  /auth/send-code:
    post:
      consumes:
      - application/json
      description: Sends a verification code to the specified email or phone number.
      parameters:
      - description: Recipient and device UID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.sendCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Send verification code
      tags:
      - verification
  /auth/verify/email:
    post:
      consumes:
      - application/json
      description: Verifies a user's email address using a code sent to them. Requires
        Bearer token.
      parameters:
      - description: Email, code and device UID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.verifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Verify email address
      tags:
      - verification
  /auth/verify/phone:
    post:
      consumes:
      - application/json
      description: Verifies a user's phone number using a code sent to them. Requires
        Bearer token.
      parameters:
      - description: Phone, code and device UID
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.verifyCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponse'
      security:
      - BearerAuth: []
      summary: Verify phone number
      tags:
      - verification
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"